requirements.txt
argcomplete==3.2.1
asttokens==2.4.1
build==1.0.3
CacheControl==0.13.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
crashtest==0.4.1
cryptography==41.0.7
decorator==5.1.1
distlib==0.3.8
docutils==0.20.1
dulwich==0.21.7
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.13.1
idna==3.6
importlib-metadata==7.0.1
installer==0.7.0
ipython==8.22.1
jaraco.classes==3.3.0
jedi==0.19.1
jeepney==0.8.0
keyring==24.3.0
markdown-it-py==3.0.0
matplotlib-inline==0.1.6
mdurl==0.1.2
more-itertools==10.1.0
msgpack==1.0.7
nh3==0.2.15
packaging==23.2
parso==0.8.3
pexpect==4.9.0
pipenv==2023.11.15
pipx==1.4.3
pkginfo==1.9.6
platformdirs==3.11.0
poetry==1.7.1
poetry-core==1.8.1
poetry-plugin-export==1.6.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.17.2
pyproject_hooks==1.0.0
rapidfuzz==3.6.1
readme-renderer==42.0
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.0
SecretStorage==3.3.3
setuptools==69.0.3
shellingham==1.5.4
six==1.16.0
stack-data==0.6.3
tk==0.1.0
tomlkit==0.12.3
traitlets==5.14.1
trove-classifiers==2024.1.8
twine==4.0.2
urllib3==2.1.0
userpath==1.9.1
virtualenv==20.25.0
wcwidth==0.2.13
wheel==0.42.0
zipp==3.17.0
